Extract Header Fragment from the List screen. #285 (#287)
