+++ eventdiscovery-sdk/src/main/java/com/schedjoules/eventdiscovery/framework/eventlist/EventListFragment.java
@@ -17,23 +17,17 @@
 
 package com.schedjoules.eventdiscovery.framework.eventlist;
 
-import android.app.Activity;
-import android.content.Intent;
-import android.content.res.Resources;
 import android.databinding.DataBindingUtil;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.design.widget.Snackbar;
-import android.support.v7.widget.Toolbar;
-import android.util.TypedValue;
 import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.TextView;
 
 import com.schedjoules.client.eventsdiscovery.Envelope;
 import com.schedjoules.client.eventsdiscovery.Event;
@@ -42,20 +36,17 @@ import com.schedjoules.client.insights.steps.Screen;
 import com.schedjoules.eventdiscovery.R;
 import com.schedjoules.eventdiscovery.databinding.SchedjoulesFragmentEventListBinding;
 import com.schedjoules.eventdiscovery.discovery.SimpleCoverageTest;
-import com.schedjoules.eventdiscovery.framework.common.BaseActivity;
 import com.schedjoules.eventdiscovery.framework.common.BaseFragment;
 import com.schedjoules.eventdiscovery.framework.common.ExternalUrlFeedbackForm;
 import com.schedjoules.eventdiscovery.framework.eventlist.view.EventListMenu;
-import com.schedjoules.eventdiscovery.framework.locationpicker.LocationPickerPlaceSelection;
-import com.schedjoules.eventdiscovery.framework.locationpicker.SharedPrefLastSelectedPlace;
 import com.schedjoules.eventdiscovery.framework.serialization.Keys;
 import com.schedjoules.eventdiscovery.framework.serialization.commons.Argument;
 import com.schedjoules.eventdiscovery.framework.serialization.commons.OptionalArgument;
 import com.schedjoules.eventdiscovery.framework.utils.InsightsTask;
+import com.schedjoules.eventdiscovery.framework.utils.fragment.Add;
 import com.schedjoules.eventdiscovery.framework.utils.fragment.ChildFragmentContainer;
 import com.schedjoules.eventdiscovery.framework.utils.fragment.FragmentContainer;
 import com.schedjoules.eventdiscovery.framework.utils.loadresult.LoadResult;
-import com.schedjoules.eventdiscovery.framework.widgets.TextWithIcon;
 
 import org.dmfs.android.microfragments.FragmentEnvironment;
 import org.dmfs.android.microfragments.utils.BooleanDovecote;
@@ -72,37 +63,20 @@ import org.dmfs.pigeonpost.localbroadcast.SerializableDovecote;
  */
 public final class EventListFragment extends BaseFragment implements EventListMenu.Listener
 {
-    private TextView mToolbarTitle;
     private EventListMenu mMenu;
     private SchedjoulesFragmentEventListBinding mViews;
     private Dovecote<Boolean> mCoverageDoveCote;
     private Dovecote<LoadResult<ResultPage<Envelope<Event>>>> mEventsLoadDoveCote;
-    private Dovecote<Boolean> mRetryDoveCote;
-    private boolean mHasOnActivityResult;
+    private Dovecote<Boolean> mReloadDovecote;
     private FragmentContainer mListFragmentContainer;
+    private boolean mIsInitializing = true;
 
 
     @Override
     public void onCreate(@Nullable Bundle savedInstanceState)
     {
         super.onCreate(savedInstanceState);
-
         new InsightsTask(getActivity()).execute(new Screen(new StringToken("list")));
-
-        mListFragmentContainer = new ChildFragmentContainer(this, R.id.schedjoules_event_list_list_holder);
-
-        if (savedInstanceState == null)
-        {
-            showResultPage(new Argument<>(Keys.EVENTS_RESULT_PAGE,
-                    new FragmentEnvironment<Bundle>(this).microFragment().parameter()).get());
-        }
-    }
-
-
-    private void showResultPage(ResultPage<Envelope<Event>> resultPage)
-    {
-        mListFragmentContainer.replace(resultPage.items().iterator().hasNext() ?
-                EventListListShowFragment.newInstance(resultPage) : EventListListNoEventsFragment.newInstance());
     }
 
 
@@ -115,7 +89,7 @@ public final class EventListFragment extends BaseFragment implements EventListMe
         mMenu = new EventListMenu(this);
         setHasOptionsMenu(true);
 
-        setupToolbar(mViews);
+        mListFragmentContainer = new ChildFragmentContainer(this, R.id.schedjoules_event_list_list_container);
 
         mCoverageDoveCote = new SerializableDovecote<>(getActivity(), "coverage-test", new Dovecote.OnPigeonReturnCallback<Boolean>()
         {
@@ -142,90 +116,42 @@ public final class EventListFragment extends BaseFragment implements EventListMe
                         }
                         else
                         {
-                            mListFragmentContainer.replace(EventListListErrorFragment.newInstance(mRetryDoveCote.cage()));
+                            mListFragmentContainer.replace(EventListListErrorFragment.newInstance(mReloadDovecote.cage()));
                         }
 
                     }
                 });
 
-        mRetryDoveCote = new BooleanDovecote(getContext(), "retry-events-load", new Dovecote.OnPigeonReturnCallback<Boolean>()
+        mReloadDovecote = new BooleanDovecote(getContext(), "reload", new Dovecote.OnPigeonReturnCallback<Boolean>()
         {
             @Override
             public void onPigeonReturn(@NonNull Boolean notUsed)
             {
-                reloadList();
+                mListFragmentContainer.replace(
+                        EventListListLoaderFragment.newInstance(
+                                new OptionalArgument<>(Keys.DATE_TIME_START_AFTER,
+                                        new FragmentEnvironment<Bundle>(EventListFragment.this).microFragment().parameter()),
+                                mEventsLoadDoveCote.cage()));
             }
         });
 
+        if (mIsInitializing)
+        {
+            new Add(R.id.schedjoules_event_list_header_container, EventListHeaderFragment.newInstance(mReloadDovecote.cage())).commit(this);
+
+            showResultPage(new Argument<>(Keys.EVENTS_RESULT_PAGE,
+                    new FragmentEnvironment<Bundle>(this).microFragment().parameter()).get());
+        }
+
+        mIsInitializing = false;
         return mViews.getRoot();
     }
 
 
-    private void setupToolbar(SchedjoulesFragmentEventListBinding views)
+    private void showResultPage(ResultPage<Envelope<Event>> resultPage)
     {
-        Toolbar toolbar = views.schedjoulesEventListToolbar;
-        toolbar.setTitle(""); // Need to set it to empty, otherwise the activity label is set automatically
-
-        mToolbarTitle = views.schedjoulesEventListToolbarTitle;
-        mToolbarTitle.setOnClickListener(new View.OnClickListener()
-        {
-            @Override
-            public void onClick(View v)
-            {
-                new LocationPickerPlaceSelection().start(EventListFragment.this);
-            }
-        });
-
-        BaseActivity activity = (BaseActivity) getActivity();
-        activity.setSupportActionBar(toolbar);
-
-        Resources res = activity.getResources();
-        toolbar.setContentInsetsAbsolute(res.getDimensionPixelSize(R.dimen.schedjoules_list_item_padding_horizontal), toolbar.getContentInsetRight());
-
-        if (res.getBoolean(R.bool.schedjoules_enableBackArrowOnEventListScreen))
-        {
-            //noinspection ConstantConditions
-            activity.getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-        }
-
-        updateToolbarTitle();
-    }
-
-
-    private void updateToolbarTitle()
-    {
-        TypedValue typedValue = new TypedValue();
-        getActivity().getTheme().resolveAttribute(R.attr.schedjoules_dropdownArrow, typedValue, true);
-        mToolbarTitle.setText(new TextWithIcon(getContext(), new SharedPrefLastSelectedPlace(getContext()).get().namedPlace().name(), typedValue.resourceId));
-    }
-
-
-    @Override
-    public void onActivityResult(int requestCode, int resultCode, Intent data)
-    {
-        mHasOnActivityResult = resultCode == Activity.RESULT_OK;
-    }
-
-
-    @Override
-    public void onResume()
-    {
-        super.onResume();
-        if (mHasOnActivityResult)
-        {
-            mHasOnActivityResult = false;
-            updateToolbarTitle();
-            reloadList();
-        }
-    }
-
-
-    private void reloadList()
-    {
-        mListFragmentContainer.replace(
-                EventListListLoaderFragment.newInstance(
-                        new OptionalArgument<>(Keys.DATE_TIME_START_AFTER, new FragmentEnvironment<Bundle>(this).microFragment().parameter()),
-                        mEventsLoadDoveCote.cage()));
+        mListFragmentContainer.replace(resultPage.items().iterator().hasNext() ?
+                EventListListShowFragment.newInstance(resultPage) : EventListListNoEventsFragment.newInstance());
     }
 
 
@@ -274,7 +200,7 @@ public final class EventListFragment extends BaseFragment implements EventListMe
     {
         mCoverageDoveCote.dispose();
         mEventsLoadDoveCote.dispose();
-        mRetryDoveCote.dispose();
+        mReloadDovecote.dispose();
         super.onDestroyView();
     }
