+++ eventdiscovery-sdk/src/main/res/layout/schedjoules_fragment_event_list_header.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout>
+
+    <android.support.v7.widget.Toolbar xmlns:android="http://schemas.android.com/apk/res/android"
+            android:id="@+id/schedjoules_event_list_toolbar"
+            android:layout_width="match_parent"
+            android:layout_height="?attr/actionBarSize">
+
+        <TextView
+                android:id="@+id/schedjoules_event_list_toolbar_title"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:maxLines="1"
+                android:ellipsize="middle"
+                style="@style/SchedJoules_ToolbarTitle"/>
+
+    </android.support.v7.widget.Toolbar>
+</layout>
